Fix potential large memory allocation in DerInputStream (#43)
* Fix potential large memory allocation in DerInputStream
The object length limit check was adding 2 int values which could result in an overflow exception, allowing very large lengths to bypass the limit enforcement.
Now the limit check casts the value to a long before doing addition to avoid integer overflow.

+++ crypto-conditions/src/main/java/com/ripple/cryptoconditions/der/DerInputStream.java
@@ -103,7 +103,7 @@ public class DerInputStream extends FilterInputStream {
     DerObject obj = new DerObject();
     obj.setTag(readTag(innerBytesRead));
     obj.setLength(readLength(innerBytesRead));
-    if (innerBytesRead.get() + obj.getLength() > limit) {
+    if ((long) innerBytesRead.get() + obj.getLength() > limit) {
       throw new DerEncodingException(
           "Object length [" + obj.getLength() + "] is larger than allowed.");
     }
